refactor(cache): split cache_test.go into focused test files (#187)
