+++ internal/cache/cache_keys_test.go
@@ -1,0 +1,87 @@
+package cache
+
+import (
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/devrecon/ludus/internal/config"
+)
+
+func TestEngineKey_Deterministic(t *testing.T) {
+	cfg := &config.Config{
+		Engine: config.EngineConfig{
+			SourcePath:      "/fake/path",
+			Version:         "5.7.3",
+			MaxJobs:         8,
+			Backend:         "native",
+			DockerBaseImage: "ubuntu:22.04",
+		},
+	}
+
+	k1 := EngineKey(cfg)
+	k2 := EngineKey(cfg)
+	if k1 != k2 {
+		t.Error("EngineKey should be deterministic for same config")
+	}
+	if k1 == "" {
+		t.Error("EngineKey should return non-empty string")
+	}
+}
+
+func TestGameServerKey_DifferentArchDifferentKey(t *testing.T) {
+	cfg := &config.Config{
+		Engine: config.EngineConfig{
+			SourcePath: "/fake/engine",
+			Version:    "5.7.3",
+		},
+		Game: config.GameConfig{
+			ProjectPath:  "/fake/project.uproject",
+			ProjectName:  "TestGame",
+			ServerTarget: "TestGameServer",
+			GameTarget:   "TestGame",
+			ServerMap:    "/Game/Maps/TestMap",
+			Arch:         "amd64",
+		},
+	}
+
+	keyAmd64 := GameServerKey(cfg, "abc123")
+	cfg.Game.Arch = "arm64"
+	keyArm64 := GameServerKey(cfg, "abc123")
+
+	if keyAmd64 == keyArm64 {
+		t.Error("GameServerKey should produce different keys for different architectures")
+	}
+	if keyAmd64 == "" || keyArm64 == "" {
+		t.Error("GameServerKey should return non-empty strings")
+	}
+}
+
+func TestContainerKey_DifferentPort(t *testing.T) {
+	tmpDir := t.TempDir()
+	if err := os.WriteFile(filepath.Join(tmpDir, "test.bin"), []byte("data"), 0644); err != nil {
+		t.Fatal(err)
+	}
+
+	cfg := &config.Config{
+		Game: config.GameConfig{
+			ProjectName:  "TestGame",
+			ServerTarget: "TestGameServer",
+		},
+		Container: config.ContainerConfig{
+			ServerPort: 7777,
+			Tag:        "latest",
+		},
+	}
+
+	key1 := ContainerKey(cfg, tmpDir)
+	cfg.Container.ServerPort = 8888
+	key2 := ContainerKey(cfg, tmpDir)
+
+	if key1 == key2 {
+		t.Error("ContainerKey should produce different keys for different ports")
+	}
+	if key1 == "" || key2 == "" {
+		t.Error("ContainerKey should return non-empty strings")
+	}
+}
